criado metodo put, getid, delete, post

+++ src/main/java/com/example/cardapio/food/FoodResponseDto.java
@@ -1,6 +1,6 @@
 package com.example.cardapio.food;
 
-public record FoodResponseDto(Long id, String title, String image, Integer price) {
+public record FoodResponseDto(Long id, String title, String image, Double price) {
     public FoodResponseDto(Food food) {
         this(food.getId(), food.getTitle(), food.getImage(), food.getPrice());
     }
